Revert "Actually merge from master"
This reverts commit 056d6fbe4f384386f3c35d0f142a1cf9a6dd532a.

+++ libraries/fbx/src/FBXReader_Node.cpp
@@ -54,8 +54,7 @@ template<class T> QVariant readBinaryArray(QDataStream& in, int& position) {
             in.readRawData(compressed.data() + sizeof(quint32), compressedLength);
             position += compressedLength;
             arrayData = qUncompress(compressed);
-            if (arrayData.isEmpty() ||
-                (unsigned int)arrayData.size() != (sizeof(T) * arrayLength)) { // answers empty byte array if corrupt
+            if (arrayData.isEmpty() || arrayData.size() != (sizeof(T) * arrayLength)) { // answers empty byte array if corrupt
                 throw QString("corrupt fbx file");
             }
         } else {
